endpoint: Limit proxy completion timeout to proxy updates
Wait for proxy redirect completion before generating policies that
redirect traffic to proxy redirect ports, to prevent traffic loss.
Reduce the proxy completion timeout to 10 seconds again, since
policy regeneration is now excluded from the timeout.

Fixes: #2788
